requirements.txt
alembic==1.16.5
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Limiter==4.1.1
psycopg[binary]==3.2.13
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
gunicorn==23.0.0
python-dotenv==1.0.1
openpyxl==3.1.5
resend
Werkzeug==3.1.6
